Timeticks are unsiged! Thanks to Josh Wilmes <josh-ucd@hitchhiker.org>
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/branches/V4-2-patches@5485 06827809-a52a-0410-b366-d66718629ded

+++ apps/snmptrapd_log.c
@@ -428,9 +428,9 @@ static void handle_time_fmt (char * bfr,
 
   /* handle output in Unix time format */
   if (fmt_cmd == CHR_CUR_TIME)
-    sprintf (safe_bfr, "%ld", (long) time_val);
+    sprintf (safe_bfr, "%lu", time_val);
   else if (fmt_cmd == CHR_UP_TIME && !options->alt_format)
-    sprintf (safe_bfr, "%ld", (long) time_val);
+    sprintf (safe_bfr, "%lu", time_val);
   else if (fmt_cmd == CHR_UP_TIME) {
     int centisecs, seconds, minutes, hours, days;
